Adds CODE OF CONDUCT in MARKDOWN (GFM)
Adds CODE OF CONDUCT in GITHUB FLAVOURED MARKDOWN

Adds community standards appropriate to this open letter. This commit is based on jculpon and peri@punchelvis.org's current understanding of proper community standards as incorporated in the USENIX Foundation and NetBSD foundation's codes of conduct and statements on Black Lives Matter and ending Asian, Asian-American, Pacific Islander, and other forms of discrimination in the United States of America. Nothing continained in this commit consists of legal advice or pastoral care as understood by the Society of Friends or the Theosophist Society of Arizona.

Signed-off-by: Jamie Culpon <peri@foxontheinter.net>